Try a hack

+++ audio_light_show/src/jukebox.rs
@@ -42,7 +42,7 @@ pub fn blink_lights(time_freqs: Vec<TimeFreq>) {
         thread::sleep(Duration::from_secs_f32(time - secs));
         secs = time;
 
-        let pin_to_light = note_to_led(freq_to_note(freq), PINS.len());
+        let pin_to_light = note_to_led(freq_to_note(freq) as i8, PINS.len());
 
         info!("Pin to light: {}", pin_to_light);
         match lit_pin {
